Pull out name, username, email mentions
Not particularly concerned about them being public, but in an effort
to make this repository more of a template for people I filled in
those fields with labels similar to the passwords.

Added TwitchTV to irc configuration, remove the scripts directory
from the path and pull in the one useful script I had there with
new instructions to "install" it, update my homebrew installed list,
and update the README to reflect these changes.